Fixes for tracing
